+++ src/services/rhai/telegram.rs
@@ -9,17 +9,17 @@ pub fn register_functions(engine: &mut Engine) {
     engine.register_result_fn(
         "send_message",
         |this: &mut Telegram, unique_id: i64, text: &str, options: Map| -> FnResult {
-            Ok(Dynamic::from(
-                this.call::<TelegramMessage>(
-                    &TelegramMethodCall::SendMessage {
-                        chat_id: TelegramChatId::UniqueId(unique_id),
-                        text: text.into(),
-                        parse_mode: get_option(&options, "parse_mode"),
-                    },
-                    None,
-                )
-                .map_err(to_string)?,
-            ))
+            this.call::<TelegramMessage>(
+                &TelegramMethodCall::SendMessage {
+                    chat_id: TelegramChatId::UniqueId(unique_id),
+                    text: text.into(),
+                    parse_mode: get_option(&options, "parse_mode"),
+                },
+                None,
+            )
+            .map(Dynamic::from)
+            .map_err(|error| error.to_string())
+            .map_err(Into::into)
         },
     );
 
@@ -27,18 +27,18 @@ pub fn register_functions(engine: &mut Engine) {
         "send_video",
         |this: &mut Telegram, unique_id: i64, bytes: Arc<Bytes>, options: Map| -> FnResult {
             // TODO: `options` parameter.
-            Ok(Dynamic::from(
-                this.call::<TelegramMessage>(
-                    &TelegramMethodCall::SendVideo {
-                        chat_id: TelegramChatId::UniqueId(unique_id),
-                        video: None, // TODO: override for `&str`.
-                        caption: get_option(&options, "caption"),
-                        parse_mode: get_option(&options, "parse_mode"),
-                    },
-                    Some(("video".into(), bytes)),
-                )
-                .map_err(to_string)?,
-            ))
+            this.call::<TelegramMessage>(
+                &TelegramMethodCall::SendVideo {
+                    chat_id: TelegramChatId::UniqueId(unique_id),
+                    video: None, // TODO: override for `&str`.
+                    caption: get_option(&options, "caption"),
+                    parse_mode: get_option(&options, "parse_mode"),
+                },
+                Some(("video".into(), bytes)),
+            )
+            .map(Dynamic::from)
+            .map_err(|error| error.to_string())
+            .map_err(Into::into)
         },
     );
 }